请求参数与注入基本完，接下来是做缓存和异常处理

+++ router-annoations/src/main/java/router/injector/InjectByteExtra.java
@@ -11,5 +11,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.CLASS)
 public @interface InjectByteExtra {
     String key();
-    int defaultValue();
+    byte defaultValue();
 }
